Organizes Areas, Shops, Bounties, and Characters
Adds /Area, /Shop, /Bounty, /Character
