+++ src/scripts/components/Act/Card.js
