+++ src/scripts/components/Act/ManagePage.js
