+++ src/scripts/components/Act/RemoveModal.js
